+++ x/marker/keeper/msg_server.go
@@ -2,8 +2,6 @@ package keeper
 
 import (
 	"context"
-	"fmt"
-	"regexp"
 
 	"github.com/armon/go-metrics"
 
@@ -40,14 +38,9 @@ func (k msgServer) AddMarker(goCtx context.Context, msg *types.MsgAddMarkerReque
 		return nil, sdkerrors.Wrap(sdkerrors.ErrInvalidRequest, "a marker can not be created in an ACTIVE status")
 	}
 
-	params := k.GetParams(ctx)
-
 	// Add marker requests must pass extra validation for denom (in addition to regular coin validation expression)
-	valExp := params.UnrestrictedDenomRegex
-	if v, expErr := regexp.Compile(valExp); expErr != nil {
-		return nil, expErr
-	} else if !v.MatchString(msg.Amount.Denom) {
-		return nil, fmt.Errorf("invalid denom (fails unrestricted marker denom validation %s): %s", valExp, msg.Amount.Denom)
+	if err = k.ValidateUnrestictedDenom(ctx, msg.Amount.Denom); err != nil {
+		return nil, err
 	}
 
 	addr := types.MustGetMarkerAddress(msg.Amount.Denom)
@@ -70,7 +63,7 @@ func (k msgServer) AddMarker(goCtx context.Context, msg *types.MsgAddMarkerReque
 		msg.MarkerType)
 	ma.SupplyFixed = msg.SupplyFixed
 
-	if params.EnableGovernance {
+	if k.GetEnableGovernance(ctx) {
 		ma.AllowGovernanceControl = true
 	} else {
 		ma.AllowGovernanceControl = msg.AllowGovernanceControl