+++ CHANGELOG.md
@@ -41,10 +41,12 @@ Ref: https://keepachangelog.com/en/1.0.0/
 ### Bug Fixes
 
 * Wired recovery flag into `init` command #254
+* Always anchor unrestricted denom validation expressions #258
 
 ### Client Breaking
 
 * Update marker transfer request signing behavior #246
+
 
 ## [v1.1.1](https://github.com/provenance-io/provenance/releases/tag/v1.1.1) - 2021-04-15